Display type of ::first-line pseudo-elements
According to the spec, the display property should not be applied
to ::first-line or ::first-letter pseudo-elements.
https://www.w3.org/TR/css-display-3/#placement

This has already been handled for ::first-letter, so this change
similarly adjusts the display property style for ::first-line. I've
tested the behavior in FireFox and legacy Edge, and this change
aligns with the behavior in other browsers.

Bug: 685925
Change-Id: Idaa299c90e20c69902491837b53e7988da7759c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2278321
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Alison Maher <almaher@microsoft.com>
Cr-Original-Commit-Position: refs/heads/master@{#784655}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1ba67ca4014ed86cdf03f0d42e7fa1edc193bff1