Remove dependency of paint offset translation on compositing
Before this CL, some non-direct compositing reasons induced a
paint offset translation for descendant reasons - in particular:
  CompositingReason::kComboCompositedDescendants
  CompositingReason::kNegativeZIndexChildren
  CompositingReason::kRoot

and we depended on the computation of these reasons in
CompositingRequirementsUpdater.

With this CL, we only do so for kComboCompositedDescendants, and
remove the dependency on CompositingRequirementsUpdater.

Bug: 1100711

Change-Id: I8b9948cfae01b0973623aba84a1fdf84ae0a9611
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2276521
Commit-Queue: Chris Harrelson <chrishtr@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#784695}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6b69ef4cead11c4bff330389d7d288934f4cb198