[content_shell --run-web-tests] Ensure main frame focus in Mac headless mode
In Mac headless mode we skips RenderWidgetHostView for main frame
activation/focus because it doesn't work without the native window.
However, previously we ensure focus of the main frame at the beginning of
the test only. When the test then navigates to a cross-domain URL, we
create new RenderWidgetHost and view, and RenderFrameHostManager
transfers focus status from the old RenderWidgetHostView to the new
RenderWidhgetHostView, which doesn't work in Mac headless mode. This
caused a web test depending on focus status to fail if it ran after
a test in a different domain.

Now after navigation (in Shell::DidNavigateMainFramePostCommit()
implementing WebContentsDelegate), ensure the activation/focus of the
main frame if it has been activated/focused before the navigation.

Bug: 1096149
Change-Id: Ic2c73901e98a268f27aecb27b066070a1afc2fdc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2269548
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: danakj <danakj@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#784956}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0f021a88d30b77281b2b04f5f744c1fb24f91cd7