Add ToString() method to DelegatedInkMetadata and improve logging
I didn't add a ToString() method to the DelegatedInkMetadata class when
I first made it, and have decided that was a mistake. Adding that method
now, and going back to use it on all event traces so that I can more
easily follow metadata as it flows through the execution.

Bug: 1052145
Change-Id: I38f17877f4fcd935cb83f08d875706941a9986c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2275154
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Daniel Libby <dlibby@microsoft.com>
Commit-Queue: Mario Bianucci <mabian@microsoft.com>
Cr-Original-Commit-Position: refs/heads/master@{#785391}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2c66086e2431676f9fa0a3636157c00de1856c4c