Remove ignore_root_layer_flings
Browser-side fling has shipped in chrome, and webview has switched to
the same implementation (though differ slightly in trigger path and
conditions). ignore_root_layer_flings was added at a time when webview
was driving fling from the browser side, but chrome was driving fling on
the renderer, and it tells the renderer side to not fling the root
layer.

So it is no longer needed, and the conditions break css snap points.

Bug: 1097648
Change-Id: I5ed29788d68eb3315dced47d5eec53cc06a46790
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2276228
Reviewed-by: David Bokan <bokan@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Commit-Queue: Bo <boliu@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#785440}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 07adea6eb0994a1989357014d06b57e01d7eb997