Add support for evaluating foldables media queries and environment vars
Add support for the screen-spanning media feature evaluation along with
the fold-* environment variables. When the window segments are
propagated into Blink, ensure that media queries are re-evaluated for
the local frame subtree. Additionally, set the environment variables on
each of the documents. The fold environment variables and
screen-spanning media feature values are both derived from the window
segments, the former being the rectangle between the segments, and the
latter determining whether the segments are side-by-side or stacked on
top of each other.


Bug: 1039051
Change-Id: Ida00c1f415486fbff3a1ad0d2269871279688691
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2262194
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Cr-Original-Commit-Position: refs/heads/master@{#785483}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: da9e208df96ace7f9e4a02f179fb8387cf865d38