Speculative fix for crbug.com/1101967
This most likely regeressed in
https://chromium-review.googlesource.com/c/chromium/src/+/2258955

In that CL, I moved the implementation of NavigateLocalAdsFrame() from
document.cc to oom_intervention_impl.cc. The only behavior change I can
see is that instead of giving FrameLoadRequest an origin context in its
constructor, I switched it to nullptr.

The crash reports aren't providing terribly useful stack traces, but my
guess is that this is crashing at
https://chromium.googlesource.com/chromium/src/+/962a6c5c187940f445c958c627875113ca31297a/content/renderer/render_frame_impl.cc#5685

RequestorOrigin() is DCHECKed non-null at the start of the function, but
I believe it would be null if FrameLoadRequest was given a nullptr
origin context.

Bug: 1101967
Change-Id: I573e8c13adc8d26fa10644f77c4b609ebfe0ccfd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2283257
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Nate Chapin <japhet@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#785783}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a94391437cac11b0f4123703adc2e24ecb610e56