[run_web_tests.py] Keep stderr of the last failure for repeated tests
This is useful for debugging flaky tests with --iterations=n (or
repeat-each=n) and --exit-after-n-failures=1. Previously the first stderr
was kept which might be for a pass which was confusing in results.html
where the failure results and the passing stderr were displayed.

This doesn't affect normal runs (e.g. on bots) where tests are not
repeated in one run (not including retries).

Also output stderr for timeout tests, which is also useful for debugging.

Change-Id: I25c1296e14b22763f36eb96d4028896f28a11c36
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2284215
Reviewed-by: Robert Ma <robertma@chromium.org>
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#785827}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 603161d59e792be28fbfb74f4e3ffd62d9a6eb73