[XProto] Fix buffer overflow when reading a RandR Notify event
When reading unions, the raw bytes are memcpy()ed into the destination.
This works for XCB where structures are usually laid out as they would
appear on the wire.  However, XProto makes some changes (padding bytes
are hidden, length fields are hidden when possible, etc), so this
approach will not work.

This CL changes RandR NotifyEvent to use a switch instead of a union.
Since we have the switch variable available, there's no reason not to.

Furthermore, there are only 5 uses of union in all of xcbproto: the one
removed here, one in xproto (which I've checked is safe), and 3 in Xkb.
All of them have switch variables, so we can change all of them over
and get rid of unions altogether, which I'll investigate in followups.

This fixes a buffer overflow when resizing a VM while Chrome is running,
but also exposes a second issue: we're calling XRRUpdateConfiguration()
without having initialized Xlib's Xrandr, so the Xlib event is nullptr
causing a crash.  I'll investigate a fix for that in a followup.

R=nickdiego

Bug: 1102059
Change-Id: Ieac8807257eab6a90a9eab26b192f7444ddfa0c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2283940
Commit-Queue: Thomas Anderson <thomasanderson@chromium.org>
Reviewed-by: Nick Yamane <nickdiego@igalia.com>
Cr-Original-Commit-Position: refs/heads/master@{#785890}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 93c83306cd7ecce655d1703e6971d7dfd3c0cafd