[Paint Preview] Avoid crash in capturing local frames separately
As I understood it, a local frame needed to be loaded to reach the point
where an embedding token was requested in
LocalFrameView::CapturePaintPreview(). However, canary crash reports
suggest it is possible a token may gone at this point (either the frame
is unloaded or had yet to finish navigating). As such, we should check
for the token and if it is not present capture should proceed as if the
frame could not be captured separately (this appears to be a very rare
case).

Bug: 1103157
Change-Id: I8a2d244cdf5e453807ef9a229a55671895a2231e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2287811
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#786305}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 24165090fc958f21bd1262cf7d26c86305532db1