Make system color keywords compute to themselves
Prior to this change, system color keywords would be resolved to RGBA
values at computed-value time. With this change, we defer resolution
until resolved-value time, to align with css-color-4.

The kernel of the change is in StyleBuilderConverter which allows for
storing an unresolved system color keyword, and in StyleColor which can
now be constructed from a system color keyword and requires a
WebColorScheme value to resolve colors. The remainder of the change
consists of updates at call sites to plumb in the used color scheme.

Bug: 1081945
Change-Id: I3b704cdf38c72b5284ad61a20e28825ee3e8e9e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2253198
Commit-Queue: Kevin Babbitt <kbabbitt@microsoft.com>
Reviewed-by: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Cr-Original-Commit-Position: refs/heads/master@{#786380}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e6e31d2c53493fd52291fffbbce7ebdcc52d8d5a