Move MarkFirstContentfulPaint call from SVGRootPainter to specific painters.
This ensures that it does not get called when not actually painting,
which enables a failing paint-timing test to pass.

The test is here:
https://source.chromium.org/chromium/chromium/src/+/master:third_party/blink/web_tests/external/wpt/paint-timing/fcp-only/fcp-svg.html

In JavaScript, the circle with id "circle" is changed from display:none
to display:block, and the test asserts that First Contentful Paint only
occurs after the change to display:block. Before this CL, the test fails
because SVGRootPainter calls MarkFirstContentfulPaint while the circle,
which is the only thing to actually be painted, is still display:none.

Moving the MarkFirstContentfulPaint calls into the individual types
(image, text, shape, foreign object) fixes this bug and aligns the
first contentful paint implementation more with largest contentful paint.


Bug: 1071450
Change-Id: I3d2a4549e00501057b4a6cf14cd04bd8f20d9d89
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2285532
Reviewed-by: Philip Rogers <pdr@chromium.org>
Commit-Queue: Annie Sullivan <sullivan@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#786421}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fe5c7a114d6b194ee70ef3f39b8c9efbb704f35d