End each web test with a single renderer-browser control ipc.
Now that renderer local dumps are all synchronous operations (thanks to
https://chromium-review.googlesource.com/c/chromium/src/+/2284373) we
can compute all the renderer-side dump results and pass them to the
WebTestControlHost in the browser immediately in InitiateCaptureDump()
ipc.

Once we do this, we no longer need to call a separate CaptureDump() ipc
to the renderer, store a callback there, and reply asynchronously. The
browser already has everything it needs once it finishes doing text
(aka layout) and pixel dumps.

We also can combine the InitiateLayoutDump() ipc with the
InitiateCaptureDump() so that we only have one ipc to the browser when
the renderer's main frame wants to signal that the test is done. To do
this we just need to move some code around in the renderer side in order
to land on a single InitiateCaptureDump() call.

This greatly simplifies the end of the test, and removes a bunch of
async states from both the browser and the renderer.

We change expectations for 2 web tests that do weird things with the
main window's WebContents. The portal test activates and destroys the
WebContents out from under WebTestControlHost, which it doesn't like.
Now it should end the test gracefully. The web platform test does a
window.close() which closes the Shell in the middle of the test, which
again the WebTestControlHost is not a fan of. But it should also now
print a FAIL indicating this happened and end the test instead of timing
out.

R=avi@chromium.org

Bug: 866140, 1069111, 1039247
Change-Id: I051a18f875dff758232cc1ba3f573f0ab4bcde2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2285169
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: danakj <danakj@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#786433}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f293c6060f9b11c72c0602c19cd5f464c50b8e69