Finalize XRSpace.NativeFromMojo
All XRSpaces, except for XRReferenceSpaces, implement MojoFromNative and
their NativeFromMojo method is simply a wrapper to invert the other
method. Because the StageParameters's standing_transform property was
floor_from_mojo instead of mojo_from_floor, XRReferenceSpaces used
NativeFromMojo and MojoFromNative was the inverter.

This change renames (and inverts) StageParameters's standing_transform
to mojo_from_floor. This removes the need for child classes of XRSpace
to implement NativeFromMojo. XRSpace still implements this as a
(non-virtual) helper method, because some math operations require it,
and this provides a convenience wrapper around the inversion.

As noted, this does invert the existing meaning of standing transform,
so the runtimes needed to be updated.

Fixed: 1031598
Change-Id: Ia36fe21ea2c179acb4a0d1a7096c7e5271d7a3a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2278317
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Klaus Weidner <klausw@chromium.org>
Reviewed-by: Brandon Jones <bajones@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#786454}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 204a824d669a7e82ad6bac6525bd73c0f1f3fab0