Fix MacOS focus ring colors pre-10.14
Based on the output I found from the new focus ring pixel browsertests I
am working on, I found that the focus ring colors on MacOS are different
between MacOS 10.14/10.15 and 10.13/earlier.
Based on a discussion with Mason and Ionel we determined that the 10.14
and later focus ring color is desired, and this patch makes the 10.13
and earlier versions have the same focus ring color.

This is tested via the new focus ring tests I am still working on. You
can see the difference with and without this patch by looking at the
test results on the different MacOS trybots in this patch at patch sets
7 and 8:
https://chromium-review.googlesource.com/c/chromium/src/+/2150765/7
https://chromium-review.googlesource.com/c/chromium/src/+/2150765/8

I could include this patch with the tests, but they got reverted in the
past and I don't want this behavior to go back and forth just because
the focus ring tests are too brittle.

Bug: 1103300
Change-Id: I483560cdad7894c25c2137389123373de4ff1935
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2288350
Reviewed-by: Ionel Popescu <iopopesc@microsoft.com>
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#786769}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 775c4d237e1a0205f2cf0eb46ffb32bee8dc62d0