[LayoutNG] Only apply block-end margins where the block actually ends.
If we reached the end of a block in an earlier fragmentainer, but
there's overflowing content, we'll revisit this block in subsequent
fragmentainers as long as child content needs it. Repeating the
block-end margins in these fragmentainers would be wrong, though.

Use AdjustForFragmentation() only for margins, and rename it to
AdjustMarginsForFragmentation().
The fieldset layout algorithm was using this one for borders/padding,
while block and multicol layout called
AdjustBorderScrollbarPaddingForFragmentation() in container_builder_.
Become more uniform by calling that method for all in
SetupFragmentBuilderForFragmentation() (via the base layout algorithm
constructor). Since this depends on the initial fragment geometry being
set, also do that in the base layout algorithm constructor.

Bug: 829028
Change-Id: I6808d023e4e79ab489bdb3a48e3c6c47cf9db6c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2287313
Reviewed-by: Alison Maher <almaher@microsoft.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#786791}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 16d77059618ccd66aca5236639bb0101f7f98bda