Blink Clipboard: Separated the plain_text from the WriteHTML function.
WriteHTML wrote both an html and a text payload. This is unnecessary
because WriteHTML ended up reimplementing much of the logic in
WriteText. Split the writing of the text and html payloads and added a
check to the WritePlainText function. Replaced all occurrences of
WriteHTML with a WriteHTML and WritePlainText call as necessary.

Change-Id: Ifccbdb9018ae945ed9bd6f51c000a7ac13733a50
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2252950
Commit-Queue: Dylan Sleeper <dsleeps@google.com>
Reviewed-by: Darwin Huang <huangdarwin@chromium.org>
Reviewed-by: enne <enne@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#786828}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e225ef11134b3c30f8cd460c743fcb600fb2f5fa