Fix compositor page-based scrolling
In https://crrev.com/7bf3fbf5532b7 we implemented page-based granularity
scrolling in the compositor scrolling path. However, this path isn't
used in production yet. There was a small bug in that CL where we
convert the scroll delta to a percent but then call
ResolveScrollPercentageToPixels with the originally passed in
|scroll_delta| rather than the modified |pixel_delta|.

This is caught by web tests like
fast/events/platform-wheelevent-paging-x-in-scrolling-page.html but this
test suite currently routes all input to the main thread, hence why it
wasn't caught. This CL also adds a unit test to make this easier to
catch.

This CL also fixes an inconsistency between main and compositor threads
in that the viewport rect used to calculate the amount of scroll
includes scrollbars on the main thread but excludes them in the
compositor. This CL makes it so that we always exclude scrollbars and
fixes the wheelevent-paging test to verify this behavior.

Bug: 1082589
Change-Id: Ic8723c4d80750570f823899518190b0ff680cff0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2281320
Commit-Queue: David Bokan <bokan@chromium.org>
Reviewed-by: Rahul Arakeri <arakeri@microsoft.com>
Cr-Original-Commit-Position: refs/heads/master@{#786846}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bb8abfeca78b0528bb1e88fd5a54a79affccb9c3