Move AddressSpace to ExecutionContext
It's never propagated to RemoteFrames, so it doesn't strictly need to
be on SecurityContext. It's debatable whether it belongs logically on
ExecutionContext or SecurityContext, but it leads to more succinct code
if it's on ExecutionContext.

Change-Id: Iafe0cb95923596d8f030b4b187504a0237dcb137
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2285531
Auto-Submit: Nate Chapin <japhet@chromium.org>
Commit-Queue: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#787038}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 50e9ad6d2febae160b3f599473d339d2f92ba764