FormControlsRefresh is now the default - move tests out of virtual [1]
This is one CL of many, moving tests out of the virtual/controls-refresh
directory, into fast/forms. This CL moves:

 virtual/controls-refresh/color -> fast/forms/color

This is a pure move of the above folder(s). A few tests share names
in the old/new location, but in all cases, the version being
moved is updated/new and should overwrite the old version.

See also crbug.com/1102647 - there is additional cleanup that will
be needed after this move is complete. But to keep the review process
simple, this is just a pure move.

Bug: 1034611
Change-Id: I05c5d682bffa686982d350839bd6d2ff1594764f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2284017
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Commit-Queue: Mason Freed <masonfreed@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#787052}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4a2ade8170862a35209cb32005d0fa095b482923