Add OnStartLoadingBody() method to WorkerMainScriptLoaderClient
When using WorkerMainScriptLoader to load the top-level script for
dedicated worker or shared worker, it's better to pass the resource
response to it's client before starting to load the body. Thus, the
client could do necessary validation in time and cancel to load the
body in case of any invalidation.

This CL adds OnStartLoadingBody() method to WorkerMainScriptLoaderClient
to achieve this aim.

Bug: 860403
Change-Id: If545611fed23b9ccd9e6602d79553604da09be24
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2277482
Reviewed-by: Makoto Shimazu <shimazu@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Minggang Wang <minggang.wang@intel.com>
Cr-Original-Commit-Position: refs/heads/master@{#787092}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 93f238135dc0dcc9a7a43a986995b93ca16c6dbf