ManifestParserTest: Tweak test to show a case that will change behavior.
This test shows a manifest with no start URL, whose document URL is
within the given scope. The test is changing to show the document URL
being in a subdirectory of the scope, not just a file in the same
directory.

This is critical to demonstrate that the expected output will change in
an upcoming CL, as the scope will, in the future, become invalidated and
default to the document URL's parent directory.

Bug: 1101909
Change-Id: Ifba108fad15148b8a7c9de6499d376d98026a178
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2290484
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Commit-Queue: Matt Giuca <mgiuca@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#787112}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2c83897fc5c427c3fbb9801dac382e1b46955395