Construct StyleResolver in StyleEngine constructor.
Years ago, the StyleResolver was deleted and re-created for certain
style sheet changes. The current lifetime of the StyleResolver is that it
is created once on first EnsureStyleResolver and only freed when the
StyleEngine is disposed. Instead of checking if it needs to be
constructed every time we access it, create it as part of constructing
the StyleEngine.

We don't create it for documents without a Frame for which we should not
use the StyleResolver.

Change-Id: I1f1d2ef322d2599737f584191550f769dcf6828e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2270209
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Robert Flack <flackr@chromium.org>
Reviewed-by: Majid Valipour <majidvp@chromium.org>
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#787158}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dc63d9714044c5cc7825ed1d511d7810ba93a9a8