Apply adopted sheet changes in DidMutate().
We only did this for DidMutateRules() and not for changes for properties
on the CSSStyleSheet object like 'disabled' and 'media'.

Unified DidMutate and DidMutate rules into a method with mutation type
parameter. We only need to invalidate the matched properties cache when
we change rules.

Simplified the code a bit for adoptedSheets which can only be set on
scopes in the same document where the adopted sheet was created.

Bug: 1102341
Change-Id: Ia996c532cdb4d90a1506008a13183b0a067f87be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2283204
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#787190}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8734154bae4101e4911b8c9178d432286c636c20