Remove key modifiers from DropData and WebDragData
While key modifiers are used and stored in DataObject, they were only
present in WebDragData and DropData to support BrowserPlugin code that
was removed in https://crrev.com/c/2287300.

After this change, the modifiers are only passed as a parameter in drag
and drop methods when needed. Their type was changed to uint32_t in some
methods to unify it with the methods already migrated to Mojo.

Bug: 1039255
Change-Id: I7fadfa371834ab862d4f4587ee6b77e50de5e36c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2289425
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Henrique Ferreiro <hferreiro@igalia.com>
Cr-Original-Commit-Position: refs/heads/master@{#787196}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9df691ebf3ab59bf9c7bc76b3d9261287e2d42cd