Use constructor document for IsConstructed().
The spec[1] term is that the constructor document is set to the
associated document for the global object. Rename the
associated_document_ to constructor_document_ to make it clearer this is
about constructed stylesheets. The constructor_document_ is always non-
null for constructed stylesheets and always null for non-constructed
stylesheets. Remove the boolean is_constructed_ which doesn't add
anything extra.

@import is not allowed in constructed stylesheets per spec, but if we
did support it, there was a case where we would not return the
constructor_document_ as OwnerDocument() for @imports. Fix that so that
it won't be a problem if we introduce @import support.

[1] https://wicg.github.io/construct-stylesheets/#constructing-stylesheets

Change-Id: I1d0a8ba0a21f48c7ec5f70b2831794c8a26ea020
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2282757
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Xiaocheng Hu <xiaochengh@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#787225}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 051963435bb7af71ab65141b3f981c5373d74324