Extend network CSP parser to support full CSP syntax
The network Content Security Policy parser is currently used only for
a few directives, while most of the Content Security Policy logic is
being handled in blink where the CSP headers are being parsed again.

This change extends the network parser to be able to parse the full
CSP grammar. We will need this parser in order to parse the csp iframe
attribute for Content Security Policy Embedded Enforcement (this is
now being done in blink, but we want to move the check to the
AncestorThrottle).

Change-Id: I638b9c5ef6f69483ab9b4f43b4454b6349180d1a
Bug: 1094909
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2220053
Reviewed-by: Camille Lamy <clamy@chromium.org>
Reviewed-by: Mike West <mkwst@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Commit-Queue: Antonio Sartori <antoniosartori@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#787241}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ff826e1f05aca3ea91a316b9488aaf76b651029e