[privacy_budget] Flush metrics at document shutdown.
For testing purposes there needs to be a predictable point at which
metrics collected on behalf of a document can be expected to be flushed
to the browser process.

This CL adds that synchronization point at the transition of the
document from active to 'stopping'

It also adds an end-to-end browser test exercising the identifiability
metrics collection logic.

R=caraitto@chromium.org, masonfreed@chromium.org

Bug: 973801
Change-Id: Ifd735ff2429271ed6fc46dc3921dcb3596c862e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2275479
Commit-Queue: Asanka Herath <asanka@chromium.org>
Reviewed-by: Mason Freed <masonfreed@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#787261}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 58c0e34211dde00f5804f8c111fffdbc5d935ee6