Fix OOPIF compositing rect calculation for multicolumn
MapAncestorToLocal and MapLocalToAncestor do not work for multicolumn
unless they are mapping a point that lies within the boundaries of
the local object. When mapping a rect rather than a point, the center
of the rect is used as the implicit mapped point. But when computing
the compositing rect for an iframe, we start with the viewport rect of
the local root view, and the center point of that rect does not
necessarily lie within the iframe's bounds.

This patch sidesteps the problem by generating a transformation matrix
from the iframe to the local root view and then inverting it. Because
we start in the coordinate space of the iframe; and because the
particular invocation uses an implicit mapped point of (0, 0) in the
iframe's coordinate space, we satisfy the requirement that the mapped
point must lie within the bounds of the local object.

This change affects some rounding behavior, which exposes some
numerical precision issues in the ScaledIframeRasterSize test. This
patch makes the arithmetic in the test more like the arithmetic in the
render process; and it correctly accounts for the 2px border on the
iframe.

BUG=1098779

Change-Id: I4b1f445da88ddb3e00b54450333072ce8354749f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2285176
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Commit-Queue: Stefan Zager <szager@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#787395}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6d5ef08ba6b0eee501105f5c8354c7ea23defb58