heap, wtf: Cleaning up ListHastSetNode
This CL deals with 2 issues in ListHashSetNode:

Tracing
=======
ListHashSetNode::Trace traces the node's value and the next_/prev_
nodes. The next_/prev_ nodes are written to atomically but do not
have a write barrier. That means that if we rely on these references
to retain the nodes, we can have races in incremental marking.

In practice, nodes in the set are retained by the backing HashMap.
Tracing a ListHashSet traces the entire HashMap. Tracing an iterator
to the ListHashSet triggers tracing of the entire ListHashSet.

This CL removes the tracing of next_/prev_ from ListHashSetNode::Trace
to indicate that nodes are not retained by other nodes, but only by the
HashMap (which has write barriers and is safe from races).

Finalization
============
Garbage collected nodes are reclaimed by the sweeper when no longer
reachable. During reclamation, the node destructor is called. Calling
the destructor manually from Destroy means that the destructor of
garbage collected nodes is called twice. This didn't result in errors
because ListHashSetNode maintains a boolean flag denoting whether it
was already destructed or not. This is a bit weird because we are
accessing a field of a destructed object (although not yet reclaimed).

We can resolve this by simply not eagerly destroying garbage collected
objects. As a bonus, the branch checking whether an object is garbage
collected can be optimized away at compile time, while the existing
branch (which checks a field) could not.

Additionally, ListHashSetNode inherits from ConditionalDestructor so
that it gets a destructor if the Type used in the set is garbage
collected and not trivially destructible. Since HeapListHashSet only
supports Member types (which are trivially destructible), the condition
will always be false and inheriting from ConditionalDestructor is
redundant.

Bug: 986235
Change-Id: Ib7ca79f79b831e097e65d05970d0ba12f5a0c493
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2292301
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Omer Katz <omerkatz@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#787663}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ee6af85d26bf09f6ae299e9fb462a9ae56b389d8