Loader: Remove legacy IPC messages for mixed contents and certificate errors
To remove the following legacy IPC messages, this CL introduces
ContentSecurityNotifier mojom interface.

  - FrameHostMsg_DidRunInsecureContent
  - FrameHostMsg_DidDisplayContentWithCertificateErrors
  - FrameHostMsg_DidRunContentWithCertificateErrors

This new interface (and the legacy IPC messages) is used for that a
renderer process notifies the browser process of errors related to mixed
contents and certificate errors.

[BEFORE THIS CL]

The legacy IPC messages were sent from RenderFrameImpl for frames and
from WebWorkerFetchContext for dedicated workers to WebContents. These
were not sent for shared worker and service workers because the messages
were routed to the ancestor render frame host but those workers do not
have the concept of the ancestor render frame.

[AFTER THIS CL]

The new IPC messages were routed via ContentSecurityNotifier mojom
interface. This interface is used only when an error happens so it is
lazily created.

For frames, ContentSecurityNotifier is created and owned by
DocumentLoader. For keeping consistency with workers (see below), this
could be done by FrameFetchContext, but it is not feasible because the
notifier can be used for navigation before context creation.

For workers, ContentSecurityNotifier is created and owned by
WorkerFetchContext. The notifier is created for shared workers and
service workers, but the pending receiver is simply ignored in the
browser process for the reasons mentioned above. See also comments in
BrowserInterfaceBinder.

Bug: 1093940
Change-Id: I527c7da08bda24aedac53dfcef41a9f1e42ee8a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2274818
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Commit-Queue: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#787691}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 761b9b0230392ef09f2ceb09a0276d52ee5e7f30