[LayoutNG] Insert OOF nodes in existing fragmentainer
Because the OOF-positioned elements are laid out after the
fragmentainers in a fragmentation context, we need a way to insert them
inside of an already laid out fragmentainer. This CL introduces a
fragmentainer "mutable copy" to which we can append NGLayoutResults
without a relayout. Once we append the resulting fragment to the
"mutable copy", we replace the original by this one in the
NGContainerFragmentBuilder.

Bug: 1079031
Change-Id: I40610943f591575ab631f24d6a2b3f9e363a46af
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2252927
Commit-Queue: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Morten Stenshorne <mstensho@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#787771}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: e2bd191dbc13ef91a5a4521f7846b5d260dd13d9