Separate DocumentInit and SecurityContextInit
Several pieces of state (SecurityOrigin, ContentSecurityPolicy, and
WebSandboxFlags) are set or calculated on DocumentInit, then transferred
to SecurityContextInit, which is then passed to SecurityContext to
complete initialization.

All of this state comes from DocumentLoader (for main thread cases).
Have DocumentLoader calculate the final state directly, then either
pass it to SecurityContextInit (in the case of SecurityOrigin) or apply
it directly to the SecurityContext for (CSP and sandbox flags).

Change-Id: I40a0a6a7f68a043818eeb9eb2ccbab437e8448d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2293175
Commit-Queue: Nate Chapin <japhet@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#787825}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ee22074eb69afa6e799262918e05c6d1e4397e5c