Protect access to end_time_ via tryLock
HandleStoppableSourceNodes() reads from end_time_ on the audio thread,
but end_time_ can be modified on the main thread, when stop() is
called.

So add a tryLock on process_lock_ here to protect access of end_time_.
stop() already establishes a lock on process_lock_ so we're all set.

Manually tested the fix for 1104573 and no races after 15 minutes.

For 1104676, the race no longer happens, but we now see a race after
15-30 sec in OfflineAudioDestinationHandler::Uninitialize(), but that's
not the issue in 1104676.

Bug: 1104573, 1104676
Change-Id: I9b6291d0f3c4085497e22545b14df1ef817689d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2295694
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Commit-Queue: Raymond Toy <rtoy@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#787965}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 82667939aebf505eed6aca8f7b534d29aedc8b7a