Add a flag for experimenting with memory measurement
If the flag is enabled then a single memory measurement is performed
for the main frame after a random time interval. The result of the
measurement is discarded after construction, so the whole measurement
is not observable to the application.

The motivation is to get stability and performance coverage for the API.

Bug: 1105494
Change-Id: Ie4500ddf31f7877ccb6e593c1319e9b1f026d8c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2297959
Commit-Queue: Ulan Degenbaev <ulan@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#788310}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 13d709cd102ea06ffd3f3c53c7acf0c8586b1094