Fix inspector-protocol/stylesheet-tracking-restart.js
The test became flaky because it was relying on the order in which
stylesheets are created on the backend. This CL changes the test to
rely on sourceURL instead of stylesheetId (incremental) to make the
test not flaky. Drive-by: adds .cache created by clangd to .gitignore

Fixed: 1104288
Change-Id: I243d97018c2153d228f78e07bd55b625265ec8da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2299219
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#788538}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ad7c7c6be0be583770a40452885be4abdc27dd3b